+++ tests/js/realm-tests.js
@@ -207,6 +207,21 @@ module.exports = {
         TestCase.assertEqual(realm.readOnly, true);
     },
 
+    testRealmOpen: function() {
+        let realm = new Realm({schema: [schemas.TestObject], schemaVersion: 1});
+        realm.write(() => {
+            realm.create('TestObject', [1])
+        });
+        realm.close();
+
+        return Realm.open({schema: [schemas.TestObject], schemaVersion: 2}).then(realm => {
+            const objects = realm.objects('TestObject');
+            TestCase.assertEqual(objects.length, 1);
+            TestCase.assertEqual(objects[0].doubleCol, 1.0);
+            realm.close();
+        });
+    },
+
     testDefaultPath: function() {
         const defaultPath = Realm.defaultPath;
         let defaultRealm = new Realm({schema: []});
@@ -1181,4 +1196,12 @@ module.exports = {
             new Realm({schema: schema, deleteRealmIfMigrationNeeded: true, migration: function(oldRealm, newRealm) {}});
         }, "Cannot include 'migration' when 'deleteRealmIfMigrationNeeded' is set.")
     },
+
+    testDisableFileFormatUpgrade: function() {
+        Realm.copyBundledRealmFiles();
+
+        TestCase.assertThrowsContaining(() => { 
+            new Realm({ path: 'dates-v3.realm', disableFormatUpgrade: true } );
+        }, 'The Realm file format must be allowed to be upgraded in order to proceed.');
+    }
 };